pyproject.toml
[project]
name = "modulate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.5"
dependencies = ["resonate-sdk>=0.4.10"]
